requirements.txt
discord.py[voice]              # The discord library
python-dotenv